refactor: 변경된 api 함수 및 타입 적용
