+++ src/components/apply/FileField.tsx
@@ -4,11 +4,14 @@ import FileItem from '@/components/common/file/FileItem';
 import InputFile from '@/components/common/input/InputFile';
 import Title from '@/components/common/title/Title';
 import { APPLY_MESSAGE } from '@/constants/applyMessages';
-import useCreatePresignedUrlsQuery from '@/hooks/useCreatePresignedUrlsQuery';
+import useCreatePresignedUrlsQuery from '@/hooks/useCreatePresignedUrlsMutation';
 import { useToastActions } from '@/stores/toastStore';
-import { NewPortfolio, PortfolioResponse } from '@/types/apis/answer';
-import { Question } from '@/types/apis/question';
-import { PresignedUrlResponse } from '@/types/apis/uploadFile';
+import {
+  NewPortfolio,
+  PortfolioResponse,
+  PresignedUrlResponse,
+  Question,
+} from '@/types/apis/application';
 import { validateMaxSize } from '@/utils/validateFileMaxSize';
 import { splitValidAndInvalidFiles } from '@/utils/validateInvalidFile';
 
@@ -64,7 +67,7 @@ function FileField({ data, onChange, values }: FileFieldProps) {
   const [portfolios, setPortfolios] = useState<NewPortfolio[]>(formatDraftValues(values) ?? []);
   const [invalidFiles, setInvalidFiles] = useState<File[]>([]);
   const [totalSize, setTotalSize] = useState(0);
-  const { createPresignedUrlsMutate } = useCreatePresignedUrlsQuery();
+  const { mutate: createPresignedUrlsMutate } = useCreatePresignedUrlsQuery();
   const { addToast } = useToastActions();
 
   const addFile = async (newFiles: FileList | null) => {
