+++ src/components/apply/UrlField.tsx
@@ -4,7 +4,7 @@ import InputField from '../common/input/InputField';
 import Title from '../common/title/Title';
 
 import { APPLY_MESSAGE } from '@/constants/applyMessages';
-import { Question } from '@/types/apis/question';
+import { Question } from '@/types/apis/application';
 import { validateUrlDetail, validateUrlStartHttp } from '@/utils/validateUrl';
 
 interface UrlFieldProps {
