+++ src/components/apply/selectBox.tsx
@@ -4,7 +4,7 @@ import Icon from '@/components/common/icon/Icon';
 import InputField from '@/components/common/input/InputField';
 import { Select } from '@/components/common/select/Select';
 import useCloseOutside from '@/hooks/useCloseOutside';
-import { JobFamily } from '@/types/apis/question';
+import { JobFamily } from '@/types/apis/application';
 
 interface selectBoxProps {
   selectedJob: JobFamily | null;
